scripts/vrs_pixmatch/requirements.txt
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
fonttools==4.47.2
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
pillow==10.2.0
pixelmatch==0.3.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
python-dateutil==2.8.2
six==1.16.0
typing==3.7.4.3
vrs==1.0.4
zipp==3.17.0
